Year switching seems to work fine now
I found that HFL.html used iDefaultYear which was hard coded as a global variable, not getting the value from the config file or xml file. Now, I added a method to get the year from Settings.xml. Also had to ensure that the menu is generated after we get the year (so we know when to stop generating options in the list).

Other changes include making sure the yahoo URL the user provides in the AddSeason interface has a '/' at the end, since that's what other files expect. Also converting the year we're getting from the server to an int (not sure why this hasn't been a problem yet...)