Added support for adding and editing a season's info
Now a user can also edit a season's data, not just add one. It uses the new year as a dividing mark between a new season and editing the current one. Also noticed that Add.html used a default year variable, so I had to change that over to Settings.xml. This should be fully functional and will be uploaded to the server.