Documentation and decimals
I forgot to commit the documentation after completing it, that was a long time ago.

Also in this commit is the switch from integer scores to decimal scores, which happened at the start of the season. Now, during the season, some bugs have been uncovered in the total/average process which I have corrected.